Refactor site API
