+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "tplink_omada_client"
-version = "1.0.2"
+version = "1.1.0"
 authors = [
   { name="Mark Godwin", email="author@example.com" },
 ]
